Add docstring info and failure callback function to templates.
